test lib: change "error" to "BUG" as appropriate
Change two uses of "error" in test-lib-functions.sh to "BUG".

In the first instance in "test_cmp_rev" the author of the "BUG"
function added in [1] had another in-flight patch adding this in [2],
and the two were never consolidated.

In the second case in "test_atexit" added in [3] that we could have
instead used "BUG" appears to have been missed.

1. 0bcc30961a7 (tests: send "bug in the test script" errors to the
   script's stderr, 2018-11-19)

2. 82f8aa680e9 (test-lib-functions: make 'test_cmp_rev' more
   informative on failure, 2018-11-19)

3. 5bb8b144864 (test-lib: introduce 'test_atexit', 2019-03-13)

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>